AccountEase/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
colorama==0.4.6
crispy-bootstrap5==2025.6
distlib==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
filelock==3.19.1
iniconfig==2.1.0
inputimeout==1.0.4
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.3.8
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.2.1
qrcode==8.2
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
gunicorn==21.2.0
whitenoise==6.6.0
pymysql==1.1.1
cryptography>=41.0.0
